Add functions for pulling the data
This commit introduces functions for fetching the desired data
(builds, jobs, projects and pipelines) from the Zuul API.
It is a first part of the puller module implementation,
allowing to get the data.

The next step would be to produce a set of results to display.
Probably some periodically updated cache would be necessary.