Run update in puller periodically
This commit introduces a piece of code that causes puller
to run in an infinite loop and update the database entries
about every 6 hours.

It also moves the database session creation directly where
it is needed, so the connection is not kept open without
any actions for too long time.