Refactor for multiple views
This commit introduces multiple improvements to the whole project.
It changes the shape of config file, allowing now to produce multiple
views for jobs statuses. The puller and server parts were modified
accordingly to respect config changes. It was also necessary
to introduce new template file, so the existing one was split
into base and index to reuse as much as possible.

The code works for my current use cases, though I identified
multiple parts where further improvements shall be involved.
For example, the queries to Zuul API shall be cached, so in case
the same build is fetched from multiple views, result is reused.
The hardcoded value for discarding too old results shall be
parametrised in configuration file.