Optimize JavaScript code
In existing implementation, the code that executes
on the page load event can take a long time on large
results pages. After profiling, I discovered that this
is because there is a code that becomes executed
for every table – and the number of tables grew largely
after the annotations feature introduced. However,
the problematic code was only required for collapsible
main tables, hence the introduced change that limits
the number of objects affected by this JS code.